src/third_party/WebKit/LayoutTests/virtual/sxg-origin-trial/http/tests/loading/sxg/fallback-to-another-sxg-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Fallback redirect to another SignedHTTPExchange
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
frame "sxg_iframe" - didReceiveTitle: 
frame "sxg_iframe" - didStartProvisionalLoadForFrame
frame "sxg_iframe" - didFailProvisionalLoadWithError
This is a testharness.js-based test.
FAIL Fallback redirect to another SignedHTTPExchange promise_test: Unhandled rejection with value: "timeout"
Harness: the test ran to completion.
